// front-end/job-portal/src/app/model/job.model.ts
export interface Job {
    id: number;
    title: string;
    companyName: string;
    jobType: string;
    location: string;
    experience: string;
    minSalary: number;
    maxSalary: number;
    category: string;
    deadline: string;
    companyImageUrl: string;
    jobDescription: string;
    cvDownloadUrl: string;
    userId: number;
  }
  